feat: adds comments
