--- src/components/shared/search-button/index.js
@@ -1,3 +1,0 @@
-import SearchButton from './search-button';
-
-export default SearchButton;
