--- src/components/shared/search-modal/index.js
@@ -1,3 +1,0 @@
-import SearchModal from './search-modal';
-
-export default SearchModal;
