+++ src/components/shared/search/footer.jsx
