+++ src/components/shared/search/hits-in-index.jsx
